backend/api/requirements.txt
aiofiles==22.1.0
arq==0.25.0
asyncpg==0.27.0
boto3==1.24.89
ftfy==6.1.1
numpy==1.23.3
opencv-python-headless==4.5.4.60
patchify==0.2.3
Pillow==10.1.0
pydash==5.1.1
pyjwt==2.6.0
python-dotenv==0.21.0
python-json-logger==2.0.4
requests==2.28.1
sanic==23.6.0
Sanic_Cors==2.2.0
# SQLAlchemy==2.0.0b1 # https://github.com/sqlalchemy/sqlalchemy/issues/8631
sqlalchemy==1.4.43
transformers==4.35.1
torch==2.1.0
watchdog==2.2.0